Refactoring of init and destroy methods and app startup in tests
