--- hermes-consumers/src/main/java/pl/allegro/tech/hermes/consumers/config/HermesConsumersConfiguration.java
@@ -1,43 +1,0 @@
-package pl.allegro.tech.hermes.consumers.config;
-
-import org.springframework.context.ApplicationContext;
-import org.springframework.context.ConfigurableApplicationContext;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.support.GenericApplicationContext;
-import pl.allegro.tech.hermes.consumers.HermesConsumers;
-import pl.allegro.tech.hermes.consumers.consumer.oauth.client.OAuthClient;
-import pl.allegro.tech.hermes.consumers.consumer.rate.maxrate.MaxRateSupervisor;
-import pl.allegro.tech.hermes.consumers.consumer.sender.http.HttpClientsWorkloadReporter;
-import pl.allegro.tech.hermes.consumers.health.ConsumerHttpServer;
-import pl.allegro.tech.hermes.consumers.hooks.SpringHooksHandler;
-import pl.allegro.tech.hermes.consumers.registry.ConsumerNodesRegistry;
-import pl.allegro.tech.hermes.consumers.supervisor.monitor.ConsumersRuntimeMonitor;
-import pl.allegro.tech.hermes.consumers.supervisor.workload.ConsumerAssignmentCache;
-import pl.allegro.tech.hermes.consumers.supervisor.workload.SupervisorController;
-
-@Configuration
-public class HermesConsumersConfiguration {
-
-    @Bean
-    public ApplicationContext applicationContext() {
-        return new GenericApplicationContext();
-    }
-
-    @Bean
-    public HermesConsumers hermesConsumers(SpringHooksHandler springHooksHandler,
-                                           ConsumerHttpServer consumerHttpServer,
-                                           ConsumerNodesRegistry consumerNodesRegistry,
-                                           SupervisorController supervisorController,
-                                           MaxRateSupervisor maxRateSupervisor,
-                                           ConsumerAssignmentCache assignmentCache,
-                                           OAuthClient oAuthHttpClient,
-                                           HttpClientsWorkloadReporter httpClientsWorkloadReporter,
-                                           ConsumersRuntimeMonitor consumersRuntimeMonitor,
-                                           ConfigurableApplicationContext applicationContext) {
-        return new HermesConsumers(springHooksHandler, consumerHttpServer, consumerNodesRegistry, supervisorController,
-                maxRateSupervisor, assignmentCache, oAuthHttpClient, httpClientsWorkloadReporter,
-                consumersRuntimeMonitor, applicationContext);
-    }
-
-}
