+++ hermes-consumers/src/main/resources/application.properties
@@ -1,0 +1,1 @@
+logging.register-shutdown-hook=true
